refactor: extract _aoi_bbox to its own module, drop lazy imports
