+++ component/scripts/gee.py
@@ -11,6 +11,7 @@ from sepal_ui.scripts.gee_interface import GEEInterface
 
 from component import parameter as cp
 from component.message import cm
+from component.scripts.aoi_geometry import _aoi_bbox
 from component.scripts.assets import default_asset_id
 
 logger = logging.getLogger("SEPLAN")
@@ -53,10 +54,7 @@ async def get_layer_min_max(
     """
     # Clip to the AOI + reduce over its bounding box. aoi.geometry() would
     # dissolve the whole collection (Collection.geometry) and blow EE's 2M-edge
-    # limit for dense GAUL 2024 boundaries (e.g. Indonesia). Imported here to
-    # avoid a module-level import cycle with seplan.
-    from component.scripts.seplan import _aoi_bbox
-
+    # limit for dense GAUL 2024 boundaries (e.g. Indonesia).
     clipped = image.clip(aoi)
     reduced = clipped.reduceRegion(
         reducer=ee.Reducer.minMax(),
@@ -225,77 +223,6 @@ def get_layer(
     )
 
 
-def get_limits(
-    gee_interface: GEEInterface,
-    asset: str,
-    data_type: Literal["binary", "continuous", "categorical"],
-    aoi: Union[ee.FeatureCollection, ee.Geometry],
-    factor: int = 2,
-) -> List[int]:
-    """Computes limits or histogram keys for the given Earth Engine image based on the specified data type.
-
-    Args:
-        gee_interface (GEEInterface): The interface used to run Earth Engine operations.
-        asset (str): Google Earth Engine asset ID.
-        data_type (str): Either 'binary', 'continuous', or any other type indicating the type of data processing.
-        aoi (ee.Geometry): Area of interest.
-        factor (int, optional): Factor to multiply the nominal scale of the image. Defaults to 2 (i.e. 2x the nominal scale
-
-    Returns:
-        list: A list containing either min-max values or histogram keys depending on the data_type.
-    """
-    # We know that the treecover_with_potential asset is binary
-    if asset == default_asset_id:
-        return [0, 1]
-
-    if data_type in ["binary", "continuous"]:
-        reducer = ee.Reducer.minMax()
-
-        def get_value(reduction):
-            return list(gee_interface.get_info(reduction).values())
-
-    else:
-        reducer = ee.Reducer.frequencyHistogram()
-
-        def get_value(reduction):
-            return gee_interface.get_info(
-                ee.Dictionary(reduction.get(ee.Image(asset).bandNames().get(0))).keys(),
-            )
-
-    ee_image = ee.Image(asset).select(0)
-    # Multiply the nominal scale by 2 in case the nominal scale is finer than 45
-    scale = ee.Number(
-        ee.Algorithms.If(
-            ee_image.projection().nominalScale().lt(30),
-            ee_image.projection().nominalScale().multiply(2),
-            ee_image.projection().nominalScale(),
-        )
-    )
-
-    # If scale is less than 30, set it to 30
-    scale = ee.Algorithms.If(scale.lt(30), 30, scale)
-
-    values = get_value(
-        ee_image.reduceRegion(
-            reducer=reducer,
-            geometry=aoi,
-            scale=scale,
-            maxPixels=1e13,
-        )
-    )
-
-    logger.debug(f"get_limits_values: {values}")
-
-    # check if values are none and if so, raise a ValueError
-    if any([val is None for val in values]):
-        raise ValueError(cm.questionnaire.error.no_limits)
-
-    # depending on the scale, values from the histogram can be floats, we'll
-    # convert them to integers... sometimes we'll show more values than the
-    # asset really has
-    return sorted(set(int(float(val)) for val in values))
-
-
 async def get_limits_async(
     gee_interface: GEEInterface,
     asset: str,
@@ -352,10 +279,7 @@ async def get_limits_async(
 
     # Clip to the AOI + reduce over its bbox; geometry=aoi would dissolve the
     # whole collection (Collection.geometry) and blow EE's 2M-edge limit for
-    # dense GAUL 2024 boundaries (e.g. Indonesia). Imported here to avoid a
-    # module-level import cycle with seplan.
-    from component.scripts.seplan import _aoi_bbox
-
+    # dense GAUL 2024 boundaries (e.g. Indonesia).
     # bestEffort + a capped maxPixels coarsens the scale for big AOIs (e.g.
     # Indonesia at 30 m would take ~2 min otherwise) — an approximate range is
     # fine for the slider; small AOIs stay under the cap and remain exact.
